Moved QRCode list to home fragment
