+++ app/src/main/java/com/example/QArmy/MainActivity.java
@@ -13,34 +13,14 @@ import com.google.android.material.bottomnavigation.BottomNavigationView;
 import java.util.ArrayList;
 
 public class MainActivity extends AppCompatActivity {
-    private ArrayList<QRCode> qrCodeDataList;
-    private ListView qrCodeList;
-    private QRCodeArrayAdapter qrCodeAdapter;
-    private TextView total;
-    private TextView max;
+
 
     private BottomNavigationView bottomNavigationView;
     private ViewPager2 viewPager;
     private FragmentPagerAdapter fragmentPagerAdapter;
     private MenuItem prevMenuItem;
 
-    public void addQRCode(QRCode qrCode) {
-        qrCodeAdapter.add(qrCode);
-        qrCodeAdapter.notifyDataSetChanged();
-        updateSummaries();
-    }
 
-    public void removeQRCode(QRCode qrCode) {
-        qrCodeAdapter.remove(qrCode);
-        qrCodeAdapter.notifyDataSetChanged();
-        updateSummaries();
-    }
-
-    private void updateSummaries() {
-        //TODO: Implement this
-        // This is where we would update our total, count, min, and max
-        // after deleting or adding a QR code
-    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -48,21 +28,7 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
 
 
-        qrCodeDataList = new ArrayList<QRCode>();
 
-        qrCodeList = findViewById(R.id.qr_code_list);
-        qrCodeAdapter = new QRCodeArrayAdapter(this, qrCodeDataList);
-        qrCodeList.setAdapter(qrCodeAdapter);
-
-        total = findViewById(R.id.sum_of_scores);
-        max = findViewById(R.id.max_score);
-
-        // TODO: Get rid of this
-        // THIS IS A TEMPORARY LIST WE ARE MAKING FOR VERIFICATION
-        // We won't actually want to create NEW qr codes here, since we should check if they already exist
-        // (i.e. have been scanned by someone else) first
-        addQRCode(new QRCode("QRCode1", null, null, null));
-        addQRCode(new QRCode("QRCode2", null, null, null));
 
         bottomNavigationView = findViewById(R.id.nav_view);
         bottomNavigationView.setSelectedItemId(R.id.navigation_home);
